Cygwin: regex: NONCHAR: re-add cast to int
wint_t is unsigned int and the test checks for a negative value.  Thus,
it's optimized out by gcc.  Add the cast from commit 44caccfca2433 to
avoid this.

+++ winsup/cygwin/regex/engine.c
@@ -119,7 +119,7 @@ static states step(struct re_guts *g, sopno start, sopno stop, states bef, wint_
 #define	EOW	(BOL-5)
 #define	BADCHAR	(BOL-6)
 #define	NWBND	(BOL-7)
-#define	NONCHAR(c)	((c) <= OUT)
+#define	NONCHAR(c)	((int)(c) <= OUT)
 /* sflags */
 #define	SBOS	0x0001
 #define	SEOS	0x0002
